refactor: replace quiz screen with reusable one
