+++ lib/features/review_page/presentation/providers/elaboration/elaboration_provider.dart
@@ -1,5 +1,11 @@
+import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:easy_localization/easy_localization.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_easyloading/flutter_easyloading.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
+import 'package:u_do_note/core/error/failures.dart';
+import 'package:u_do_note/core/logger/logger.dart';
 import 'package:u_do_note/core/shared/presentation/providers/shared_provider.dart';
 import 'package:u_do_note/features/review_page/data/datasources/elaboration/elaboration_remote_datasource.dart';
 import 'package:u_do_note/features/review_page/data/models/elaboration.dart';
@@ -7,6 +13,7 @@ import 'package:u_do_note/features/review_page/data/repositories/elaboration/ela
 import 'package:u_do_note/features/review_page/domain/repositories/elaboration/elaboration_repository.dart';
 import 'package:u_do_note/features/review_page/domain/usecases/elaboration/get_elaborated_content.dart';
 import 'package:u_do_note/features/review_page/domain/usecases/elaboration/save_quiz_results.dart';
+import 'package:u_do_note/features/review_page/presentation/providers/review_screen_provider.dart';
 
 part 'elaboration_provider.g.dart';
 
@@ -49,8 +56,8 @@ class Elaboration extends _$Elaboration {
 
   /// Save the quiz results
   /// This also be used to save the remark when the user has not taken the quiz
-  Future<dynamic> saveQuizResults(String notebookId,
-      ElaborationModel elaborationModel) async {
+  Future<dynamic> saveQuizResults(
+      String notebookId, ElaborationModel elaborationModel) async {
     var saveQuizResults = ref.read(saveQuizResultsProvider);
 
     var failureOrString = await saveQuizResults(notebookId, elaborationModel);
@@ -65,5 +72,37 @@ class Elaboration extends _$Elaboration {
     var failureOrContent = await getElaboratedContent(content);
 
     return failureOrContent.fold((failure) => failure, (content) => content);
+  }
+
+  Future<void> onQuizFinish(
+      BuildContext context,
+      ElaborationModel elaborationModel,
+      List<int> selectedAnswersIndex,
+      int score) async {
+    var reviewState = ref.read(reviewScreenProvider);
+
+    var updatedElaborationModel = elaborationModel.copyWith(
+        sessionName: reviewState.getSessionTitle,
+        createdAt: Timestamp.now(),
+        score: score,
+        selectedAnswersIndex: selectedAnswersIndex);
+
+    EasyLoading.show(
+        status: 'Saving quiz results...',
+        maskType: EasyLoadingMaskType.black,
+        dismissOnTap: false);
+
+    var res = await ref
+        .read(elaborationProvider.notifier)
+        .saveQuizResults(reviewState.getNotebookId, updatedElaborationModel);
+
+    EasyLoading.dismiss();
+
+    if (!context.mounted) return;
+
+    if (res is Failure) {
+      logger.e('Failed to save quiz results: ${res.message}');
+      EasyLoading.showError(context.tr("save_quiz_e"));
+    }
   }
 }
