+++ lib/features/review_page/presentation/providers/pomodoro/pomodoro_technique_provider.dart
@@ -1,5 +1,8 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_easyloading/flutter_easyloading.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
+import 'package:u_do_note/core/error/failures.dart';
 import 'package:u_do_note/core/shared/presentation/providers/shared_provider.dart';
 import 'package:u_do_note/features/review_page/data/datasources/pomodoro/pomdoro_remote_datasource.dart';
 import 'package:u_do_note/features/review_page/data/models/pomodoro.dart';
@@ -7,6 +10,7 @@ import 'package:u_do_note/features/review_page/data/repositories/pomodoro/pomodo
 import 'package:u_do_note/features/review_page/domain/entities/pomodoro/pomodoro_state.dart';
 import 'package:u_do_note/features/review_page/domain/repositories/pomodoro/pomodoro_technique_repository.dart';
 import 'package:u_do_note/features/review_page/domain/usecases/pomodoro/save_quiz_results.dart';
+import 'package:u_do_note/features/review_page/presentation/providers/review_screen_provider.dart';
 
 part 'pomodoro_technique_provider.g.dart';
 
@@ -49,5 +53,29 @@ class Pomodoro extends _$Pomodoro {
         await saveQuizResultsUseCase(notebookId, pomodoroModel);
 
     return failureOrString.fold((failure) => failure, (res) => res);
+  }
+
+  Future<void> onQuizFinish(BuildContext context, PomodoroModel pomodoroModel,
+      List<int> selectedAnswersIndex, int score) async {
+    var reviewScreenState = ref.read(reviewScreenProvider);
+
+    var updatedPomodoroModel = pomodoroModel.copyWith(
+        score: score, selectedAnswersIndex: selectedAnswersIndex);
+
+    EasyLoading.show(
+        status: 'Saving quiz results...',
+        maskType: EasyLoadingMaskType.black,
+        dismissOnTap: false);
+
+    var res = await ref
+        .read(pomodoroProvider.notifier)
+        .saveQuizResults(reviewScreenState.notebookId!, updatedPomodoroModel);
+
+    EasyLoading.dismiss();
+
+    if (res is Failure) {
+      EasyLoading.showError("Something went wrong. Please try again later.");
+      return;
+    }
   }
 }
