+++ lib/features/review_page/presentation/providers/pq4r/pq4r_provider.dart
@@ -1,5 +1,10 @@
+import 'package:easy_localization/easy_localization.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_easyloading/flutter_easyloading.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
+import 'package:u_do_note/core/error/failures.dart';
+import 'package:u_do_note/core/logger/logger.dart';
 import 'package:u_do_note/core/shared/presentation/providers/shared_provider.dart';
 import 'package:u_do_note/features/review_page/data/datasources/pq4r/pq4r_remote_datasource.dart';
 import 'package:u_do_note/features/review_page/data/models/pq4r.dart';
@@ -44,5 +49,28 @@ class Pq4r extends _$Pq4r {
     var failureOrMessage = await saveQuizResults(pq4rModel);
 
     return failureOrMessage.fold((failure) => failure, (res) => res);
+  }
+
+  Future<void> onQuizFinish(BuildContext context, Pq4rModel pq4rModel,
+      List<int> selectedAnswersIndex, int score) async {
+    var updatedPq4rModel = pq4rModel.copyWith(
+        score: score, selectedAnswersIndex: selectedAnswersIndex);
+
+    EasyLoading.show(
+        status: 'Saving quiz results...',
+        maskType: EasyLoadingMaskType.black,
+        dismissOnTap: false);
+
+    var failureOrMessage =
+        await ref.read(pq4rProvider.notifier).saveQuizResults(updatedPq4rModel);
+
+    if (!context.mounted) return;
+
+    if (failureOrMessage is Failure) {
+      logger.e('Failed to save quiz results: ${failureOrMessage.message}');
+      EasyLoading.showError(context.tr("save_quiz_e"));
+    }
+
+    EasyLoading.dismiss();
   }
 }
